# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Simple server in order to create and mantain CGA Survey"
  version: "1.0.0"
  title: "CGA Survey"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "a@a.com"
  license:
    name: "MIT"
    url: "https://github.com/TrentoCrowdAI/CGA/blob/master/LICENSE"
host: "cga-mock-server.herokuapp.com"
basePath: "/v1"
...
schemes:
- "https"
paths:
  /projects:
    get:
      tags:
      - "project"
      summary: "Get all project"
      description: "Performed by a user in order to get all projects"
      operationId: "getProjects"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/project"
        400:
          description: "Bad request"
      x-swagger-router-controller: "Project"
    post:
      tags:
      - "project"
      summary: "Add a new project"
      description: "Performed by a user when he wants to create a new project"
      operationId: "addProject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "project"
        description: "Project object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/project_1"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/project"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "Project"
  /projects/{id}:
    get:
      tags:
      - "project"
      summary: "Get an project by the id"
      description: "Performed by users in order to get specific project"
      operationId: "getProject"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Project id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/project"
        400:
          description: "Invalid id"
        404:
          description: "Project not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Project"
    put:
      tags:
      - "project"
      summary: "Update an existing project"
      description: "Performed by a user in order to update the project"
      operationId: "updateProject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Project id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Project object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/project"
        400:
          description: "Invalid id"
        404:
          description: "Project not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Project"
    delete:
      tags:
      - "project"
      summary: "Delete a Project by id"
      description: "Performed by the user in order to delete the Project from the\
        \ system"
      operationId: "deleteProject"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Project id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/project"
        400:
          description: "Invalid id"
        404:
          description: "Project not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Project"
  /dataCollections:
    get:
      tags:
      - "dataCollection"
      summary: "Get all dataCollection"
      description: "Performed by a user in order to get all dataCollections"
      operationId: "getAllDataCollections"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/dataCollection"
        400:
          description: "Bad request"
      x-swagger-router-controller: "DataCollection"
    post:
      tags:
      - "dataCollection"
      summary: "Add a new DataCollection"
      description: "Performed by a user when he wants to create a new DataCollection"
      operationId: "addDataCollection"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "dataCollection"
        description: "DataCollection object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/dataCollection_1"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/dataCollection"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "DataCollection"
  /dataCollections/{id}:
    get:
      tags:
      - "dataCollection"
      summary: "Get an DataCollection by the id"
      description: "Performed by users in order to get specific DataCollection"
      operationId: "getDataCollection"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The DataCollection id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/dataCollection"
        400:
          description: "Invalid id"
        404:
          description: "DataCollection not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "DataCollection"
    put:
      tags:
      - "dataCollection"
      summary: "Update an existing dataCollection"
      description: "Performed by a user in order to update the dataCollection"
      operationId: "updateDataCollection"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The DataCollection id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Project object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_1"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/dataCollection"
        400:
          description: "Invalid id"
        404:
          description: "DataCollection not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "DataCollection"
    delete:
      tags:
      - "dataCollection"
      summary: "Delete a DataCollection by id"
      description: "Performed by the user in order to delete the DataCollection from\
        \ the system"
      operationId: "deleteDataCollection"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The DataCollection id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/dataCollection"
        400:
          description: "Invalid id"
        404:
          description: "DataCollection not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "DataCollection"
...
  /flowConstraints:
    get:
      tags:
      - "flowConstraint"
      summary: "Get all FlowConstraints"
      description: "Performed by a user in order to get all FlowConstraint"
      operationId: "getAllFlowConstraints"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/flowConstraint"
        400:
          description: "Bad request"
      x-swagger-router-controller: "FlowConstraint"
    post:
      tags:
      - "flowConstraint"
      summary: "Add a new FlowConstraint"
      description: "Performed by a user when he wants to create a new FlowConstraint"
      operationId: "addFlowConstraint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "flowConstraint"
        description: "FlowConstraint object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/flowConstraint_1"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/flowConstraint"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "FlowConstraint"
  /flowConstraints/{id}:
    get:
      tags:
      - "flowConstraint"
      summary: "Get a FlowConstraint by the id"
      description: "Performed by users in order to get specific FlowConstraint"
      operationId: "getFlowConstraint"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The FlowConstraint id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/flowConstraint"
        400:
          description: "Invalid id"
        404:
          description: "FlowConstraint not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "FlowConstraint"
    put:
      tags:
      - "flowConstraint"
      summary: "Update an existing FlowConstraint"
      description: "Performed by a user in order to update the FlowConstraint"
      operationId: "updateFlowConstraint"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The FlowConstraint id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "FlowConstraint object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_6"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/flowConstraint"
        400:
          description: "Invalid id"
        404:
          description: "FlowConstraint not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "FlowConstraint"
    delete:
      tags:
      - "flowConstraint"
      summary: "Delete a FlowConstraint by id"
      description: "Performed by the user in order to delete the FlowConstraint from\
        \ the system"
      operationId: "deleteFlowConstraint"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The FlowConstraint id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/flowConstraint"
        400:
          description: "Invalid id"
        404:
          description: "FlowConstraint not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "FlowConstraint"
...
  /labels:
    get:
      tags:
      - "label"
      summary: "Get all label"
      description: "Performed by a user in order to get all Label"
      operationId: "getAllLabels"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/label"
        400:
          description: "Bad request"
      x-swagger-router-controller: "Label"
    post:
      tags:
      - "label"
      summary: "Add a new Label"
      description: "Performed by a user when he wants to create a new Label"
      operationId: "addLabel"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "Label"
        description: "Label object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Label"
      responses:
        201:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/label"
        400:
          description: "Bad Request"
      x-swagger-router-controller: "Label"
  /labels/{id}:
    get:
      tags:
      - "label"
      summary: "Get a Label by the id"
      description: "Performed by users in order to get specific Label"
      operationId: "getLabel"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Label id"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/label"
        400:
          description: "Invalid id"
        404:
          description: "Label not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Label"
    put:
      tags:
      - "label"
      summary: "Update an existing Label"
      description: "Performed by a user in order to update the Label"
      operationId: "updateLabel"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Label id"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Label object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/body_9"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/label"
        400:
          description: "Invalid id"
        404:
          description: "Label not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Label"
    delete:
      tags:
      - "label"
      summary: "Delete a Label by id"
      description: "Performed by the user in order to delete the Label from the system"
      operationId: "deleteLabel"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Label id"
        required: true
        type: "integer"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/label"
        400:
          description: "Invalid id"
        404:
          description: "Label not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "Label"
...
definitions:
...
  survey:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the survey in the system"
      name:
        type: "string"
        description: "Name assigned to the survey"
      description:
        type: "string"
        description: "String that describe the survey"
      components:
        type: "array"
        description: "SurveyComponent that compose a survey"
        items:
          $ref: "#/definitions/surveyComponent"
    example:
      components:
      - name: "name"
        id: 5
        items:
        - validators:
          - id: 2
            type: "type"
            value: "value"
          - id: 2
            type: "type"
            value: "value"
          options:
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          description: "description"
          rules:
          - item: 3
            condition: "condition"
            id: 9
          - item: 3
            condition: "condition"
            id: 9
          id: 5
          type: "type"
          flowConstraints:
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        - validators:
          - id: 2
            type: "type"
            value: "value"
          - id: 2
            type: "type"
            value: "value"
          options:
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          description: "description"
          rules:
          - item: 3
            condition: "condition"
            id: 9
          - item: 3
            condition: "condition"
            id: 9
          id: 5
          type: "type"
          flowConstraints:
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
      - name: "name"
        id: 5
        items:
        - validators:
          - id: 2
            type: "type"
            value: "value"
          - id: 2
            type: "type"
            value: "value"
          options:
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          description: "description"
          rules:
          - item: 3
            condition: "condition"
            id: 9
          - item: 3
            condition: "condition"
            id: 9
          id: 5
          type: "type"
          flowConstraints:
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
        - validators:
          - id: 2
            type: "type"
            value: "value"
          - id: 2
            type: "type"
            value: "value"
          options:
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          - id: 7
            type: "type"
            value: "value"
            labels:
            - values: "values"
              language: "language"
              id: 2
            - values: "values"
              language: "language"
              id: 2
          description: "description"
          rules:
          - item: 3
            condition: "condition"
            id: 9
          - item: 3
            condition: "condition"
            id: 9
          id: 5
          type: "type"
          flowConstraints:
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          - typology: "typology"
            condition: "condition"
            id: 4
            value: "value"
          labels:
          - values: "values"
            language: "language"
            id: 2
          - values: "values"
            language: "language"
            id: 2
      name: "name"
      description: "description"
      id: 1
...
  label:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the label"
      language:
        type: "string"
        description: "Contains the language of the question"
      values:
        type: "string"
        description: "Contains the question in the language indicated in the language\
          \ field"
    example:
      values: "values"
      language: "language"
      id: 2
  flowConstraint:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the label"
      typology:
        type: "string"
        description: "Describe the type of constraint"
      condition:
        type: "string"
        description: "Condition to apply the constraint"
      value:
        type: "string"
        description: "Value of the constraint"
    example:
      typology: "typology"
      condition: "condition"
      id: 4
      value: "value"
...
  subject:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the subject in the system"
      name:
        type: "string"
        description: "Name of the subject"
      surname:
        type: "string"
        description: "Surname of the subject"
      email:
        type: "string"
        description: "Email of the subject"
      role:
        type: "string"
        description: "Role of the subject in the system"
    example:
      role: "role"
      surname: "surname"
      name: "name"
      id: 7
      email: "email"
...
  project_1:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Name assigned to the project"
      description:
        type: "string"
        description: "String that describe the project"
      collections:
        type: "array"
        description: "DataCollection that compose a project"
        items:
          $ref: "#/definitions/dataCollection"
      subjects:
        type: "array"
        description: "Subject assigned to the project"
        items:
          $ref: "#/definitions/subject"
  body:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the project in the system"
      name:
        type: "string"
        description: "Name assigned to the project"
      description:
        type: "string"
        description: "String that describe the project"
      collections:
        type: "array"
        description: "DataCollection that compose a project"
        items:
          $ref: "#/definitions/dataCollection"
      subjects:
        type: "array"
        description: "Subject assigned to the project"
        items:
          $ref: "#/definitions/subject"
  dataCollection_1:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Name assigned to the dataCollection"
      description:
        type: "string"
        description: "String that describe the dataCollection"
      surveys:
        type: "array"
        description: "Surveys that compose a dataCollection"
        items:
          $ref: "#/definitions/survey"
  body_1:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the dataCollection in the system"
      name:
        type: "string"
        description: "Name assigned to the dataCollection"
      description:
        type: "string"
        description: "String that describe the dataCollection"
      surveys:
        type: "array"
        description: "Surveys that compose a dataCollection"
        items:
          $ref: "#/definitions/survey"
...
  flowConstraint_1:
    type: "object"
    properties:
      typology:
        type: "string"
        description: "Describe the type of constraint"
      condition:
        type: "string"
        description: "Condition to apply the constraint"
      value:
        type: "string"
        description: "Value of the constraint"
  body_6:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the label"
      typology:
        type: "string"
        description: "Describe the type of constraint"
      condition:
        type: "string"
        description: "Condition to apply the constraint"
      value:
        type: "string"
        description: "Value of the constraint"
...
  Label:
    type: "object"
    properties:
      language:
        type: "string"
        description: "Contains the language of the question"
      values:
        type: "string"
        description: "Contains the question in the language indicated in the language\
          \ field"
    example:
      values: "values"
      language: "language"
  body_9:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Number that identify the label"
      language:
        type: "string"
        description: "Contains the language of the question"
      values:
        type: "string"
        description: "Contains the question in the language indicated in the language\
          \ field"
...